Move all code to the Junit test. Some code cleanup.
