+++ pom.xml
@@ -45,7 +45,7 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>4.13-SNAPSHOT</version>
+            <version>4.12</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
